requirements.txt
discord.py==2.4.0
beautifulsoup4==4.12.3
requests==2.32.3
